requirements.txt
django
django-auditlog
django-allauth
django-axes
django-admin-interface
django-bootstrap-v5
django-cors-headers
django-crispy-forms
django-debug-toolbar
django-environ
django-extensions
django-filter
django-floppyforms
django-guardian
django-haystack
django-sitetree
django-redis
django-rest-passwordreset

djangorestframework
djangorestframework-simplejwt
djangorestframework-jsonapi
drf-api-logger
drf-yasg

asgiref
mysqlclient
pytz
sqlparse
Unipath
dj-database-url
python-decouple
gunicorn
whitenoise
environ
